requirements.txt
alabaster==0.7.12
asn1crypto==1.2.0
attrs==19.3.0
Babel==2.7.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
docutils==0.15.2
filelock==3.0.12
Quart==0.11.2
idna==2.8
imagesize==1.1.0
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
packaging==19.2
Pallets-Sphinx-Themes==1.2.3
pluggy==0.13.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.5
pytz==2019.3
requests==2.22.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
toml==0.10.0
tox==3.14.0
urllib3==1.25.7
virtualenv==16.7.7
Werkzeug==1.0.0
zipp==0.6.0
